Move SaveQuestion component to its own folder, and fix imports
