+++ frontend/src/metabase/components/SaveQuestionForm/schema.ts
@@ -3,7 +3,7 @@ import * as Yup from "yup";
 import * as Errors from "metabase/lib/errors";
 
 export const SAVE_QUESTION_SCHEMA = Yup.object({
-  saveType: Yup.string(),
+  saveType: Yup.string().oneOf(["overwrite", "create"]),
   name: Yup.string().when("saveType", {
     // We don't care if the form is valid when overwrite mode is enabled,
     // as original question's data will be submitted instead of the form values
